Create BinaryTree.java
In this code, we define a TreeNode class to represent nodes in the binary tree and a BinaryTree class to perform the height calculation. The findHeight method recursively calculates the height of the binary tree by finding the maximum height of the left and right subtrees and adding 1 to account for the current level. The getHeight method is used to start the height calculation from the root of the tree.