Create BinaryTree2.java
In this code, we define a TreeNode class to represent nodes in the binary tree and a BinaryTree2 class to perform the node count calculation. The countNodes method recursively counts the nodes in the binary tree by counting the current node and then recursively counting the nodes in the left and right subtrees. The getNumberOfNodes method is used to start the node count calculation from the root of the tree.