Create BinaryTree3.java
In this code, we define a TreeNode class to represent nodes in the binary tree and a BinaryTree3 class to perform the level calculation. The findLevel method recursively traverses the tree, and when it finds the node with the given data (key), it returns the current level. If the node is not found, it returns -1. The getLevel method is used to start the level calculation from the root of the tree.