Create BinaryTree4.java
In this code, we define a TreeNode class to represent nodes in the binary tree and a BinaryTree4 class to perform the inorder tree traversal. The inOrderTraversal method recursively performs the inorder traversal by visiting the left subtree, the current node, and then the right subtree. The performInOrderTraversal method is used to start the traversal from the root of the tree.