Create CustomrQueue.java
You are tasked with simulating a queue of customers waiting for service at a bank. You need to implement a queue data structure using a linked list in Java to manage this queue. The queue should support the following operations:

a) enqueue(customerID): Add a customer to the end of the queue. Each customer is represented by a unique customerID.

b) dequeue(): Remove and return the next customer in the queue. If the queue is empty, return -1 as there are no customers.

c) size(): Get the number of customers in the queue.

Implement a Java class CustomerQueue that contains these methods and write a program to demonstrate the usage of the queue. Consider the following customer interactions:

a) Three customers arrive at the bank in the order with customer IDs 1, 2, and 3, and they join the queue accordingly.

b) Two customers are serviced in the order they joined the queue, and their customer IDs are removed from the queue.

c) A new customer with customer ID 4 arrives and joins the queue.

d) Two more customers are serviced in the order they joined the queue.

e) Determine the size of the queue after all these operations are completed.